feat: WASM targets for Lean 4 (component model and Emscripten)
Add two Emscripten-based build modes:

- LEAN_WASM_COMPONENT: standalone WebAssembly module (STANDALONE_WASM, no
  pthreads) targeting the WIT component model, for use as ZED editor
  extensions. EM_ASM inline JavaScript in io.cpp/debug.cpp is replaced with
  WASI-compatible alternatives (std::getenv, exit(1)).
- Emscripten compiler build: the Lean compiler runs under node.js and
  compiles Lean files to C (`node lean.js -c test.c test.lean`). dlsym is
  replaced with EM_ASM resolving wasm exports into the indirect function
  table; PTHREAD_POOL_SIZE=4 with LEAN_MULTI_THREAD so the task manager
  initializes; 22 libuv assertion stubs replaced with POSIX implementations;
  __wasi_random_get used for WASM randomness.

Shared runtime fixes, also applying to strict-ABI and 32-bit targets:

- 32-bit detection in lean.h/thread.cpp now checks __i386__, __wasm32__,
  and __SIZEOF_POINTER__ == 4, not just LEAN_EMSCRIPTEN.
- 8 runtime signatures fixed to match the C code generator output (unused
  world-token parameters removed, Unit parameter added).
- Pre-existing stub mismatches in uv/event_loop.cpp and uv/system.cpp fixed;
  missing <mutex> include for MULTI_THREAD=OFF; olean header fields use
  uint64_t instead of size_t for cross-arch compatibility.

tests/wasm: end-to-end shell tests (uname, getenv, getpid, cwd, hrtime,
extern C codegen) and Lean evaluation tests (eval_factorial, coe_notation,
extern_resolve) with expected outputs.

+++ stage0/src/library/module.cpp
@@ -129,7 +129,8 @@ struct olean_header {
     // 40 bytes: build githash, padded with `\0` to the right
     char githash[40];
     // address at which the beginning of the file (including header) is attempted to be mmapped
-    size_t base_addr;
+    // Use fixed 64-bit width so .olean files are cross-architecture compatible.
+    uint64_t base_addr;
     // In v3, the fixed header is followed by these length-prefixed sections:
     //   size_t   data_size                                // byte length of the compacted data
     //   compacted data                                    // `data_size` bytes
@@ -140,8 +141,7 @@ struct olean_header {
     // In v2, the compacted data starts immediately after the header (no `data_size`, no sections).
     size_t data[];
 };
-// make sure we don't have any padding bytes, which also ensures `data` is properly aligned
-static_assert(sizeof(olean_header) == 5 + 1 + 1 + 33 + 40 + sizeof(size_t), "olean_header must be packed");
+static_assert(sizeof(olean_header) == 5 + 1 + 1 + 33 + 40 + sizeof(uint64_t), "olean_header must be packed");
 
 
 // Compactor external object: wraps a live `object_compactor` for incremental compaction.
